Split right sidebar palette coverage helpers
