+++ Sources/ContentView+RightSidebarCommandPalette.swift
@@ -1,0 +1,116 @@
+import Foundation
+
+extension ContentView {
+    static func commandPaletteShortcutAction(forCommandID commandId: String) -> KeyboardShortcutSettings.Action? {
+        if let rightSidebarModeAction = commandPaletteRightSidebarModeShortcutAction(forCommandID: commandId) {
+            return rightSidebarModeAction
+        }
+
+        switch commandId {
+        case "palette.newWorkspace":
+            return .newTab
+        case "palette.newWindow":
+            return .newWindow
+        case "palette.openFolder":
+            return .openFolder
+        case "palette.reopenPreviousSession":
+            return .reopenPreviousSession
+        case "palette.newTerminalTab":
+            return .newSurface
+        case "palette.newBrowserTab":
+            return .openBrowser
+        case "palette.closeWindow":
+            return .closeWindow
+        case "palette.toggleSidebar":
+            return .toggleSidebar
+        case "palette.showNotifications":
+            return .showNotifications
+        case "palette.jumpUnread":
+            return .jumpToUnread
+        case "palette.renameTab":
+            return .renameTab
+        case "palette.renameWorkspace":
+            return .renameWorkspace
+        case "palette.editWorkspaceDescription":
+            return .editWorkspaceDescription
+        case "palette.nextWorkspace":
+            return .nextSidebarTab
+        case "palette.previousWorkspace":
+            return .prevSidebarTab
+        case "palette.nextTabInPane":
+            return .nextSurface
+        case "palette.previousTabInPane":
+            return .prevSurface
+        case "palette.browserToggleDevTools":
+            return .toggleBrowserDeveloperTools
+        case "palette.browserConsole":
+            return .showBrowserJavaScriptConsole
+        case "palette.browserReactGrab":
+            return .toggleReactGrab
+        case "palette.browserSplitRight", "palette.terminalSplitBrowserRight":
+            return .splitBrowserRight
+        case "palette.browserSplitDown", "palette.terminalSplitBrowserDown":
+            return .splitBrowserDown
+        case "palette.terminalSplitRight":
+            return .splitRight
+        case "palette.terminalSplitDown":
+            return .splitDown
+        case "palette.findInDirectory":
+            return .findInDirectory
+        case "palette.terminalFind":
+            return .find
+        case "palette.terminalFindNext":
+            return .findNext
+        case "palette.terminalFindPrevious":
+            return .findPrevious
+        case "palette.terminalHideFind":
+            return .hideFind
+        case "palette.terminalUseSelectionForFind":
+            return .useSelectionForFind
+        case "palette.toggleSplitZoom":
+            return .toggleSplitZoom
+        case "palette.triggerFlash":
+            return .triggerFlash
+        default:
+            return nil
+        }
+    }
+
+    static func commandPaletteRightSidebarModeCommandContributions() -> [CommandPaletteCommandContribution] {
+        func constant(_ value: String) -> (CommandPaletteContextSnapshot) -> String {
+            { _ in value }
+        }
+
+        return RightSidebarMode.allCases.map { mode in
+            CommandPaletteCommandContribution(
+                commandId: Self.commandPaletteRightSidebarModeCommandID(mode),
+                title: constant(mode.shortcutAction.label),
+                subtitle: constant(String(localized: "command.rightSidebarMode.subtitle", defaultValue: "Right Sidebar")),
+                keywords: ["right", "sidebar", "show", "switch", "focus", mode.rawValue]
+            )
+        }
+    }
+
+    static func commandPaletteRightSidebarModeCommandID(_ mode: RightSidebarMode) -> String {
+        switch mode {
+        case .files:
+            return "palette.showRightSidebarFiles"
+        case .find:
+            return "palette.showRightSidebarFind"
+        case .sessions:
+            return "palette.showRightSidebarSessions"
+        case .feed:
+            return "palette.showRightSidebarFeed"
+        case .dock:
+            return "palette.showRightSidebarDock"
+        }
+    }
+
+    private static func commandPaletteRightSidebarModeShortcutAction(
+        forCommandID commandID: String
+    ) -> KeyboardShortcutSettings.Action? {
+        RightSidebarMode.allCases.first { mode in
+            Self.commandPaletteRightSidebarModeCommandID(mode) == commandID
+        }?.shortcutAction
+    }
+}
