+++ cmuxTests/ShortcutAndCommandPaletteTests.swift
@@ -1135,44 +1135,6 @@ final class RightSidebarModeShortcutHintTests: XCTestCase {
         XCTAssertEqual(RightSidebarMode.dock.shortcutAction, .switchRightSidebarToDock)
     }
 
-    func testCommandPaletteIncludesActionForEachRightSidebarMode() throws {
-        let contributions = ContentView.commandPaletteRightSidebarModeCommandContributions()
-        let contributionsByID = Dictionary(uniqueKeysWithValues: contributions.map { ($0.commandId, $0) })
-        let context = ContentView.CommandPaletteContextSnapshot()
-
-        for mode in RightSidebarMode.allCases {
-            let commandID = ContentView.commandPaletteRightSidebarModeCommandID(mode)
-            let contribution = try XCTUnwrap(
-                contributionsByID[commandID],
-                "Expected command palette contribution for \(mode.rawValue)"
-            )
-
-            XCTAssertEqual(contribution.title(context), mode.shortcutAction.label)
-            XCTAssertEqual(
-                contribution.subtitle(context),
-                String(localized: "command.rightSidebarMode.subtitle", defaultValue: "Right Sidebar")
-            )
-            XCTAssertTrue(contribution.keywords.contains("right"))
-            XCTAssertTrue(contribution.keywords.contains("sidebar"))
-            XCTAssertTrue(contribution.keywords.contains(mode.rawValue))
-            XCTAssertTrue(contribution.when(context))
-            XCTAssertTrue(contribution.enablement(context))
-        }
-
-        XCTAssertEqual(contributions.count, RightSidebarMode.allCases.count)
-    }
-
-    func testCommandPaletteRightSidebarActionsUseModeShortcutActions() {
-        for mode in RightSidebarMode.allCases {
-            XCTAssertEqual(
-                ContentView.commandPaletteShortcutAction(
-                    forCommandID: ContentView.commandPaletteRightSidebarModeCommandID(mode)
-                ),
-                mode.shortcutAction
-            )
-        }
-    }
-
     func testModeShortcutUsesConfiguredBindings() {
         let customFilesShortcut = StoredShortcut(
             key: "4",
